pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robot-agent-runtime"
version = "0.1.0"
description = "Generic robot agent runtime with pluggable skills."
readme = "README.md"
requires-python = ">=3.9,<3.10"
dependencies = [
  "Pillow>=10.0,<12.0",
  "ultralytics>=8.2,<9.0",
  "opencv-python==4.9.0.80",
  "numpy<2",
  "torch==2.2.2",
  "torchvision==0.17.2",
  "openai>=1.0,<2.0",
  "websockets>=12,<16",
]

[project.scripts]
robot-agent = "backend.cli:main"
robot-agent-tracking-perception = "scripts.run_tracking_perception:main"
robot-agent-tracking-loop = "scripts.run_tracking_loop:main"
robot-agent-tracking-viewer-stream = "scripts.run_tracking_viewer_stream:main"
robot-agent-tracking-stack = "scripts.run_tracking_stack:main"

[dependency-groups]
dev = [
  "pytest>=7.0,<9.0",
]

[tool.setuptools.packages.find]
include = ["backend*", "skills*", "scripts*"]

[tool.uv]
default-groups = ["dev"]
